# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: S3MP API
  version: 1.1.0
  description: S3MP 是面向多租户的受控文件存储服务。管理接口使用租户会话 Cookie `s3mp_session`；账户登录与平台控制接口使用账户会话 Cookie `s3mp_account_session`。浏览器发起变更请求时同时提交
    `X-S3MP-CSRF`，服务端应用访问使用 API Key。错误响应统一返回稳定的 `code`、可读的 `message` 和用于排障的 `request_id`；任何会话值、API Key
    或存储凭据均不会出现在响应中。
paths:
...
  /api/v1/users:
    get:
      tags:
      - Context
      summary: 列出当前租户可见的用户账户。
      operationId: list_users
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/UserStatus'
          default: active
        description: 资源当前生命周期状态。
      - name: principal_type
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/PrincipalType'
          default: user
        description: “principal_type”字段；具体取值和约束见当前资源说明。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出当前租户可见的用户账户。
      x-permission: members.read
  /api/v1/member-candidates:
    get:
      tags:
      - Context
      summary: 搜索尚未加入当前租户的全局已注册用户候选
      operationId: list_member_candidates
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberCandidatePage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 搜索尚未加入当前租户的全局已注册用户候选；支持邮箱、姓名和工号模糊查询。
      x-permission: members.manage
  /api/v1/users/{user_id}:
    get:
      tags:
      - Context
      summary: 获取当前租户中指定用户的公开身份信息。
      operationId: get_user
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: 全局用户账户的服务端唯一标识。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取当前租户中指定用户的公开身份信息。
      x-permission: members.read
  /api/v1/members:
    get:
      tags:
      - Context
      summary: 列出当前租户的成员关系。
      operationId: list_members
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/MembershipStatus'
          default: active
        description: 资源当前生命周期状态。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出当前租户的成员关系。
      x-permission: members.read
    post:
      tags:
      - Context
      summary: 将已有用户加入当前租户，或创建受邀成员关系。
      operationId: create_member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将已有用户加入当前租户，或创建受邀成员关系。
      x-permission: members.manage
  /api/v1/members/{membership_id}:
    get:
      tags:
      - Context
      summary: 获取当前租户中指定成员关系的详情。
      operationId: get_member
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取当前租户中指定成员关系的详情。
      x-permission: members.read
    patch:
      tags:
      - Context
      summary: 更新成员状态
      operationId: update_member
      parameters:
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      - name: If-Match
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: If-Match
        description: 客户端已读取的资源 ETag；用于拒绝覆盖较新的服务端状态。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 更新成员状态；暂停、移除或到期会立即回收租户权限。
      x-permission: members.manage
  /api/v1/groups/{group_id}/members:
    get:
      tags:
      - Context
      summary: 列出指定用户组中的成员关系。
      operationId: list_group_members
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出指定用户组中的成员关系。
      x-permission: groups.read
    post:
      tags:
      - Context
      summary: 将当前租户成员加入指定用户组。
      operationId: add_group_member
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddGroupMemberBody'
      responses:
        '204':
          description: Successful Response
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将当前租户成员加入指定用户组。
      x-permission: groups.manage
  /api/v1/groups/{group_id}/members/{membership_id}:
    delete:
      tags:
      - Context
      summary: 将成员从指定用户组移除。
      operationId: remove_group_member
      parameters:
      - name: group_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Group Id
        description: 用户组的服务端唯一标识。
      - name: membership_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Membership Id
        description: 租户成员关系的服务端唯一标识。
      responses:
        '204':
          description: Successful Response
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 将成员从指定用户组移除。
      x-permission: groups.manage
...
  /api/v1/platform/accounts:
    get:
      tags:
      - Platform control plane
      summary: 分页列出平台账户目录，可按邮箱、系统号、姓名和账户状态筛选。
      operationId: list_platform_accounts
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: 本页最多返回的记录数。
      - name: query
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 320
          - type: 'null'
          title: Query
        description: “query”字段；具体取值和约束见当前资源说明。
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - active
            - disabled
            type: string
          - type: 'null'
          title: Status
        description: 资源当前生命周期状态。
      - name: include_deleted
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Deleted
        description: “include_deleted”字段；具体取值和约束见当前资源说明。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 分页列出平台账户目录，可按邮箱、系统号、姓名和账户状态筛选。
      x-permission: platform.accounts.read
  /api/v1/platform/accounts/import:
    post:
      tags:
      - Platform control plane
      summary: 按 Excel 批量创建平台账户
      operationId: import_platform_accounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountImportRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountImportResult'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 按 Excel 批量创建平台账户；已占用身份和无效行不会覆盖已有账户。
      x-permission: platform.accounts.manage
  /api/v1/platform/accounts/{user_id}:
    get:
      tags:
      - Platform control plane
      summary: 获取平台账户的安全摘要
      operationId: get_platform_account
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: 全局用户账户的服务端唯一标识。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取平台账户的安全摘要；不返回密码、会话或密钥信息。
      x-permission: platform.accounts.read
    delete:
      tags:
      - Platform control plane
      summary: 执行该公开 API 操作
      operationId: delete_platform_account
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: 全局用户账户的服务端唯一标识。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 执行该公开 API 操作；具体输入、输出和授权要求见下方说明。
...
  /api/v1/platform/accounts/{user_id}/password:
    post:
      tags:
      - Platform control plane
      summary: 重置指定活跃平台账户的密码，并撤销其全部账户和租户会话。
      operationId: reset_platform_account_password
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
        description: 全局用户账户的服务端唯一标识。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 重置指定活跃平台账户的密码，并撤销其全部账户和租户会话。
      x-permission: platform.accounts.manage
  /api/v1/platform/roles:
    get:
      tags:
      - Platform control plane
      summary: 列出平台角色及其平台权限目录。
      operationId: list_platform_roles
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: 本页最多返回的记录数。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformRolePage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出平台角色及其平台权限目录。
      x-permission: platform.roles.read
  /api/v1/platform/role-bindings:
    get:
      tags:
      - Platform control plane
      summary: 按用户分页列出有效的平台角色授权集合
      operationId: list_platform_role_bindings
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: 本页最多返回的记录数。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformRoleBindingPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 按用户分页列出有效的平台角色授权集合；每个用户仅返回一项。
      x-permission: platform.roles.read
    post:
      tags:
      - Platform roles
      summary: 为一个全局账户创建或替换平台角色集合
      operationId: grant_platform_role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlatformRoleGrant'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Grant Platform Role
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 为一个全局账户创建或替换平台角色集合；一个用户仅保留一项有效授权。
  /api/v1/platform/support-access:
    get:
      tags:
      - Platform control plane
      summary: 分页列出支持访问申请及其待审批、已批准、已撤销或已过期状态。
      operationId: list_support_access
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: 本页最多返回的记录数。
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SupportAccessStatus'
          - type: 'null'
          title: Status
        description: 资源当前生命周期状态。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportAccessPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 分页列出支持访问申请及其待审批、已批准、已撤销或已过期状态。
      x-permission: platform.support.read
    post:
      tags:
      - Platform support
      summary: 提交针对指定租户的限时支持访问申请
      operationId: request_support_access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportAccessCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Request Support Access
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 提交针对指定租户的限时支持访问申请；默认不包含文件内容权限。
  /api/v1/platform/support-access/{request_id}:
    get:
      tags:
      - Platform control plane
      summary: 获取单条支持访问申请的安全详情和授权物化记录。
      operationId: get_support_access
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
        description: 请求的服务端唯一标识，用于问题追踪。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportAccessResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取单条支持访问申请的安全详情和授权物化记录。
      x-permission: platform.support.read
    delete:
      tags:
      - Platform support
      summary: 撤销已申请或已批准的支持访问，并回收其有效租户会话。
      operationId: revoke_support_access
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Request Id
        description: 请求的服务端唯一标识，用于问题追踪。
      responses:
        '204':
          description: Successful Response
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 撤销已申请或已批准的支持访问，并回收其有效租户会话。
  /api/v1/platform/audit-events:
    get:
      tags:
      - Platform control plane
      summary: 分页检索平台控制面审计事件，可按动作、资源类型和资源标识筛选。
      operationId: list_platform_audit_events
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
        description: 本页最多返回的记录数。
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 160
          - type: 'null'
          title: Action
        description: “action”字段；具体取值和约束见当前资源说明。
      - name: resource_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Resource Type
        description: “resource_type”字段；具体取值和约束见当前资源说明。
      - name: resource_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 120
          - type: 'null'
          title: Resource Id
        description: “resource_id”字段；具体取值和约束见当前资源说明。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformAuditEventPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 分页检索平台控制面审计事件，可按动作、资源类型和资源标识筛选。
      x-permission: platform.audit.read
  /api/v1/platform/audit-events/{event_id}:
    get:
      tags:
      - Platform control plane
      summary: 获取单条平台审计事件的脱敏详情。
      operationId: get_platform_audit_event
      parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Event Id
        description: “event_id”字段；具体取值和约束见当前资源说明。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformAuditEventResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 获取单条平台审计事件的脱敏详情。
      x-permission: platform.audit.read
...
  /api/v1/permission_catalog:
    get:
      tags:
      - Authorization
      summary: 获取当前租户可配置的权限目录
      description: 获取当前租户可配置的权限目录；不包含任何 platform.* 平台权限。
      operationId: get_permission_catalog
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionCatalogResponse'
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '422': *id001
        '500': *id001
      x-permission: roles.read
  /api/v1/groups:
    get:
      tags:
      - Authorization
      summary: 列出当前租户的用户组。
      operationId: list_groups
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
        description: 不透明分页游标；仅可用于相同查询条件的下一页请求。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupPage'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 列出当前租户的用户组。
      x-permission: groups.read
    post:
      tags:
      - Authorization
      summary: 创建当前租户的用户组。
      operationId: create_group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupWrite'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '422': *id001
        '400': *id001
        '401': *id001
        '403': *id001
        '404': *id001
        '409': *id001
        '412': *id001
        '500': *id001
      description: 创建当前租户的用户组。
      x-permission: groups.manage
...
components:
  schemas:
...
    MemberCandidatePage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/MemberCandidateResponse'
          type: array
          title: Items
          description: 当前页返回的资源记录列表。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页分页游标；无更多结果时为 null。
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: MemberCandidatePage
...
    PrincipalType:
      type: string
      enum:
      - user
      - group
      - application
      title: PrincipalType
...
    UserPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserResponse'
          type: array
          title: Items
          description: 当前页返回的资源记录列表。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页分页游标；无更多结果时为 null。
      additionalProperties: false
      type: object
      required:
      - items
      - next_cursor
      title: UserPage
    UserResponse:
      properties:
        id:
          type: string
          title: Id
          description: 资源的服务端唯一标识。
        email:
          type: string
          format: email
          title: Email
          description: 用户登录邮箱地址。
        display_name:
          type: string
          title: Display Name
          description: 面向用户展示的名称。
        status:
          type: string
          enum:
          - active
          - disabled
          - deleted
          title: Status
          description: 资源当前生命周期状态。
        created_at:
          type: string
          format: date-time
          title: Created At
          description: 资源创建时间，采用 Asia/Shanghai（UTC+08:00）格式。
      additionalProperties: false
      type: object
      required:
      - id
      - email
      - display_name
      - status
      - created_at
      title: UserResponse
    UserStatus:
      type: string
      enum:
      - active
      - disabled
      - deleted
      title: UserStatus
...